refactor(library): file playlist refresh under library
refresh_provider.dart holds PlaylistRefreshState and
RefreshManagerNotifier, and its consumers are library_page, home_page,
remote_playlist_sync_provider and auto_refresh_service -- none of them
search. auto_refresh_service.dart follows it: app.dart is its only
importer and what it refreshes on a timer is playlists.

lib/services/refresh/ was a single-file directory and is now gone.
